+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Changes
 
+- Changed `se_mean_cat()` into `se_prop()`
 - Updated functions to support tidy evaluation (unquoted arguments).
 - Separated `se_mean()` into two functions: `se_mean_num()` and `se_mean_cat()`.
 - Replaced the `condition` argument with the ellipsis `...`.